Adding in value/copy, collection/concatenate, and created new 'value' namespace
