+++ lib/burner/library/collection/arrays_to_objects.rb
@@ -23,7 +23,7 @@ module Burner
       #     jobs: [
       #       {
       #         name: 'set',
-      #         type: 'b/set_value',
+      #         type: 'b/value/static',
       #         value: [
       #           [1, 'funky']
       #         ]
