+++ lib/burner/version.rb
@@ -8,5 +8,5 @@
 #
 
 module Burner
-  VERSION = '1.0.0-alpha.12'
+  VERSION = '1.0.0-alpha.13'
 end
